pcb: update 3D models to avoid absolute paths
Add "Wild" as a 3D Search Paths alias in the Configure Paths dialog of
Footprint Editor that points to the kicad-libraries-wild/3d folder, that
way I don't have to worry about hard-coded aboslute paths.

Warning: kicad saves files as CRLF on Windows, so dos2unix all the
things before committing to avoid a 100% git diff. It also seems to
re-sort a lot of the data so I manually only accepted part of the diff
created by kicad.