firmware: remove I2C (Wire) entirely
Don't need I2C for Flick, so there's no need to keep that library code
around.

I had to do heavy modifications to the Adafruit graphics and BusIO
libraries to remove all references to the I2C/Wire APIs. It's not the
prettiest code, but the compiled binary is a few kB smaller and there's
fewer runtime "spi or i2c" runtime checks in the OLED code now.