firmware: get the AC working and test PHOTO_THRESH
Yay it works, but its potentiometer is also backwards, as expected.
Change the test code into a basic graph display. Also hook up the LED to
the AC IRQ, but PWM it at 5% rather than driving at full strength
because this LED is too dang bright. (Proper fix is to solder
a different resistor to the board, but using PWM outputs is easy enough
cheating)