firmware: remove PWM stuff from demo code
Now that I have a bigger resistor (2.2k rather than 330) that doesn't
blind me at full brightness, I can drop the PWM stuff and just drive the
LED as a normal GPIO, or optionally drive it directly from the AC
peripheral.

Also switch the AC digital clock from the 32.768k crystal to 48MHz from
the CPU to minimize its latency. In the future I might derive a ~1MHz
GCLK and use that, but for now just go with GCLK0 that's already set up.