firmware: uf2 bootloader: remove python2 dependency, fix upload-bossa
'make upload-bossa' now always programs the self-update version, using
the correct flags to program to offset 0x2000.

Update gendata.py to work with python3, all I had to change was '/'
operators to '//' to get integer division (all the existing IO and
bytearray code was fine), and allow parsing the bootloader size as hex.