add skeleton for flick-flasher, the PC software component
I've decided to do this in Rust using the pixels library to get
a GPU-accelerated framebuffer, and winit to handle windowing.

Pixels is a simple bitmap framebuffer, but it uses GPU acceleration to
render and display via wgpu. That said, all my "rendering" is
a CPU-bound loop operating on a byte slice, but the display logic is
simple. It also supports toggling vsync, which could be helpful.

winit handles cross-platform window management and input events, and
supports fullscreen (exclusive or borderless window) toggling.

Current code is based on the minimal-winit example in the pixels repo,
but I heavily modified it to remove winit_input_helper, support window
resizing, and sped up the draw() function so it only uses 45% CPU
instead of 75%.

The final Flick flasher graphics code will be even simpler (just toggle
between black and white) but commit the bouncing box here for posterity.