flasher: add workaround for integrated vs discrete GPU adapter selection
When I tried to run it on my desktop on Windows, wgpu decided to try to
use the integrated graphics rather than my actual GPU. Maybe "Intel"
gets sorted before "Nvidia"?

Anyway, this is a total hack but apparently requesting an adapter with
a preference for HighPerformance causes wgpu to correctly pick the
proper GPU.

Still to investigate: based on the render times it seems like vsync
might be on even when it should be turned off (both from PixelsBuilder
and seen as performance_mode in the wgpu debug logs).